refactor(ai-playground): kebab-case component filenames
Completes the rename started on the PR1 branch: AGENTS.md:67 requires
`kebab-case.vue` filenames, the `name:` property stays PascalCase (:68).
Nothing merged depends on these names.
