+++ packages/ui/components/ai-playground/bs-ai-expertise-picker.vue
