+++ packages/ui/components/ai-playground/bs-ai-skill-picker.vue
